♻️ rename delay component to wait
