--- src/components/delay/index.ts
@@ -1,3 +1,0 @@
-import Delay from './Delay'
-
-export default Delay
