+++ src/components/wait/Wait.spec.tsx
@@ -1,8 +1,8 @@
 import { render } from '@testing-library/react'
 import React from 'react'
-import Delay from './Delay'
+import Delay from './Wait'
 
-describe('Delay Component', () => {
+describe('Wait Component', () => {
   it('should have span tag with time props', () => {
     const { container } = render(<Delay time={1000} />)
     expect(container).toBeInTheDocument()
