+++ src/components/wait/index.ts
@@ -1,0 +1,3 @@
+import Wait from './Wait'
+
+export default Wait
